requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.6
djangorestframework==3.13.1
dnspython==2.2.1
executing==0.8.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.23.1
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
toml==0.10.2
traitlets==5.3.0
wcwidth==0.2.5
